requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1

# ── Data ─────────────────────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.2
requests==2.31.0

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# ── ML ───────────────────────────────────────────────────────────────────────
scikit-learn==1.4.1.post1
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1
shap==0.45.0
joblib==1.3.2

# ── Experiment Tracking ──────────────────────────────────────────────────────
mlflow==2.11.3

# ── Monitoring ───────────────────────────────────────────────────────────────
evidently==0.4.22

# ── API ──────────────────────────────────────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.0
httpx==0.27.0

# ── Dashboard ────────────────────────────────────────────────────────────────
streamlit==1.33.0
plotly==5.21.0
